added profile page
added a profile page, just shows the username for right now.  added a button that goes there that only shows if the user is logged in.  fixed the visibility of the database form button to only show if the user is logged in and not on the database-form page.  made the search bar only show when on the database-form page.